requirements.txt
streamlit==1.49.1
pandas>=1.5.0
numpy>=1.24.0
ydata-profiling==4.16.0
plotly==6.3.0
matplotlib>=3.6.0
snowflake-connector-python==3.17.3
SQLAlchemy==2.0.43
psycopg2-binary==2.9.10
google-genai==1.38.0
openai>=1.0.0
python-dotenv==1.1.1
openpyxl>=3.0.0
xlrd>=2.0.0
networkx>=3.0.0

